requirements.txt
mysql-connector==2.2.9
PyMySQL==0.9.3
testcontainers==3.5.4